requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
mailtrap==2.0.1
requests==2.32.3
python-dotenv==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
setuptools==80.9.0
Werkzeug==3.1.3
wheel==0.45.1
psycopg[binary]
